web-ui tests: share the jsdom harness across the new component tests
