+++ plugins/web-ui/test/approval-handoff.test.ts
@@ -1,10 +1,8 @@
 import { metadata } from "./model-metadata.ts";
 import assert from "node:assert/strict";
 import { test } from "node:test";
-import { JSDOM } from "jsdom";
-import { createServer } from "vite";
-import type { Conversation } from "../src/conv-types.ts";
-import type { PendingApproval } from "../src/core-bridge.ts";
+import type { PendingApproval, SessionEntry } from "../src/core-bridge.ts";
+import { bootConversation, session, until } from "./dom-harness.ts";
 
 function deferred<T>() {
   let resolve!: (value: T) => void;
@@ -12,45 +10,11 @@ function deferred<T>() {
   return { promise, resolve };
 }
 
-async function until(check: () => boolean): Promise<void> {
-  for (let i = 0; i < 200; i++) {
-    if (check()) return;
-    await new Promise((resolve) => setTimeout(resolve, 10));
-  }
-  assert.fail("condition did not settle");
-}
-
 test("approval handoff unlocks queue and steer without losing pending decisions", async (t) => {
-  const dom = new JSDOM('<!doctype html><div id="app"></div><main id="main"></main>', {
-    url: "http://localhost/",
-  });
-  Object.defineProperty(dom.window, "matchMedia", {
-    value: () => ({ matches: false, addEventListener() {}, removeEventListener() {} }),
-  });
-  const globals = {
-    window: dom.window,
-    document: dom.window.document,
-    location: dom.window.location,
-    history: dom.window.history,
-    localStorage: dom.window.localStorage,
-    navigator: dom.window.navigator,
-    HTMLElement: dom.window.HTMLElement,
-    customElements: dom.window.customElements,
-    Node: dom.window.Node,
-    Event: dom.window.Event,
-    InputEvent: dom.window.InputEvent,
-    KeyboardEvent: dom.window.KeyboardEvent,
-    requestAnimationFrame: (callback: FrameRequestCallback) => setTimeout(() => callback(Date.now()), 0),
-    cancelAnimationFrame: clearTimeout,
-    getComputedStyle: dom.window.getComputedStyle.bind(dom.window),
-    EventSource: undefined,
-  };
-  for (const [key, value] of Object.entries(globals))
-    Object.defineProperty(globalThis, key, { configurable: true, writable: true, value });
-
   const approval: PendingApproval = { requestId: "a1", command: "echo test", reason: "requires approval" };
-  const row = { id: "s1", threadRef: "web:owner:test", scopeId: "personal:owner", title: "Test" };
-  const entries = [{ seq: 1, type: "user", createdAt: Date.now(), payload: { text: "run the command" } }];
+  const entries: SessionEntry[] = [
+    { seq: 1, type: "user", createdAt: Date.now(), payload: { text: "run the command" } },
+  ];
   let selectedModelId = "gpt-5.6-sol";
   let modelDeleted = false;
   let pending = [approval];
@@ -60,14 +24,13 @@ test("approval handoff unlocks queue and steer without losing pending decisions"
   let refreshGate: ReturnType<typeof deferred<void>> | undefined;
   let submitted = false;
   const requests: Array<{ path: string; body?: Record<string, unknown> }> = [];
-  const originalFetch = globalThis.fetch;
-  globalThis.fetch = async (input, init) => {
+  const route: typeof fetch = async (input, init) => {
     const path = String(input);
     requests.push({ path, ...(init?.body ? { body: JSON.parse(String(init.body)) } : {}) });
     if (path.includes("runtime-config")) {
       if (init?.method === "PUT") selectedModelId = JSON.parse(String(init.body)).modelId;
       return Response.json({
-        scopeId: row.scopeId,
+        scopeId: session.scopeId,
         approvedHarnesses: ["pi"],
         modelsByHarness: { pi: modelDeleted ? ["replacement-api"] : ["gpt-5.6-sol"] },
         modelCatalog: modelDeleted
@@ -95,43 +58,17 @@ test("approval handoff unlocks queue and steer without losing pending decisions"
     if (path.startsWith("/api/sessions/s1")) {
       if (submitted) await handoff.promise;
       await refreshGate?.promise;
-      return Response.json({ session: row, entries, earlierEntries: 0 });
+      return Response.json({ session, entries, earlierEntries: 0 });
     }
-    if (path === "/api/sessions") return Response.json({ sessions: [row] });
+    if (path === "/api/sessions") return Response.json({ sessions: [session] });
     if (path === "/api/contexts") return Response.json({ contexts: [] });
     throw new Error(`Unexpected request: ${path}`);
   };
-  const vite = await createServer({ server: { middlewareMode: true, hmr: false }, appType: "custom" });
-  let conv: Conversation | undefined;
+  const boot = await bootConversation({ entries, fetch: route });
+  const { conv: chat, host } = boot;
   try {
-    await vite.ssrLoadModule("/src/shell.ts");
-    const { appState } = await vite.ssrLoadModule("/src/shell-state.ts");
-    const { sessionsState } = await vite.ssrLoadModule("/src/sessions.ts");
-    const { createConversation } = await vite.ssrLoadModule("/src/conversations.ts");
-    const { entriesToMessages, attachPendingApprovals } = await vite.ssrLoadModule("/src/core-bridge.ts");
-    const { transcriptModel } = await vite.ssrLoadModule("/src/model-options.ts");
-    const { seedRuntimeConfig } = await vite.ssrLoadModule("/src/composer.ts");
-    seedRuntimeConfig(row.scopeId, await (await fetch("/api/runtime-config")).json());
-    appState.me = { user: "owner", org: "test" };
-    appState.currentView = "chats";
-    sessionsState.list = [row];
-    const host = document.querySelector<HTMLElement>("#main")!;
-    appState.mainEl = host;
-    conv = createConversation({
-      pane: true,
-      ownsUrl: false,
-      container: () => host,
-      claimContainer: () => host,
-      visible: () => true,
-      density: () => "full",
-      onDensityChange() {},
-      ensureDeliveryStream() {},
-    }) as Conversation;
-    const chat = conv;
     function mount() {
-      const messages = entriesToMessages(entries, transcriptModel());
-      attachPendingApprovals(messages, pending, transcriptModel());
-      chat.mountContinuable(row.threadRef, row.id, row.scopeId, messages);
+      boot.mount(pending);
       chat.state.agent!.convertToLlm = () => [{ role: "user", content: "run the command", timestamp: 0 }];
     }
     function click(label: string) {
@@ -171,8 +108,8 @@ test("approval handoff unlocks queue and steer without losing pending decisions"
       assert.deepEqual(requests.find((r) => r.path === "/api/runs/r1/signal")?.body, {
         kind: "steer",
         text: "use the smaller change",
-        threadRef: row.threadRef,
-        scopeId: row.scopeId,
+        threadRef: session.threadRef,
+        scopeId: session.scopeId,
       });
       assert.equal(chat.state.agent!.state.isStreaming, true);
     });
@@ -273,7 +210,7 @@ test("approval handoff unlocks queue and steer without losing pending decisions"
       selectedModelId = "deleted-overlay";
       modelDeleted = true;
       mount();
-      await chat.composer.refreshRuntimeSelection(row.scopeId, chat.state.agent!);
+      await chat.composer.refreshRuntimeSelection(session.scopeId, chat.state.agent!);
       await until(() => !!host.querySelector('select[aria-label="Replacement model"]'));
       assert.equal(chat.composer.currentModelOption(), undefined);
       assert.match(host.textContent ?? "", /deleted-overlay/);
@@ -302,12 +239,6 @@ test("approval handoff unlocks queue and steer without losing pending decisions"
     refreshGate?.resolve();
     decision.resolve(Response.json({ error: "test complete" }, { status: 503 }));
     continuation.resolve(Response.json({ status: "done", result: { status: "ok", reply: "done" } }));
-    conv?.state.agent?.abort();
-    await conv?.state.agent?.waitForIdle();
-    conv?.composer.dispose();
-    conv?.dispose();
-    await vite.close();
-    globalThis.fetch = originalFetch;
-    dom.window.close();
+    await boot.dispose();
   }
 });
