+++ plugins/web-ui/test/tool-chips.test.ts
@@ -1,46 +1,10 @@
-import { metadata } from "./model-metadata.ts";
 import assert from "node:assert/strict";
 import { test } from "node:test";
-import { JSDOM } from "jsdom";
-import { createServer } from "vite";
-import type { Conversation } from "../src/conv-types.ts";
-
-async function until(check: () => boolean): Promise<void> {
-  for (let i = 0; i < 200; i++) {
-    if (check()) return;
-    await new Promise((resolve) => setTimeout(resolve, 10));
-  }
-  assert.fail("condition did not settle");
-}
+import type { SessionEntry } from "../src/core-bridge.ts";
+import { bootConversation, until } from "./dom-harness.ts";
 
 test("tool rows render their label beside a mono detail chip, tinted by outcome", async () => {
-  const dom = new JSDOM('<!doctype html><div id="app"></div><main id="main"></main>', { url: "http://localhost/" });
-  Object.defineProperty(dom.window, "matchMedia", {
-    value: () => ({ matches: false, addEventListener() {}, removeEventListener() {} }),
-  });
-  const globals = {
-    window: dom.window,
-    document: dom.window.document,
-    location: dom.window.location,
-    history: dom.window.history,
-    localStorage: dom.window.localStorage,
-    navigator: dom.window.navigator,
-    HTMLElement: dom.window.HTMLElement,
-    customElements: dom.window.customElements,
-    Node: dom.window.Node,
-    Event: dom.window.Event,
-    InputEvent: dom.window.InputEvent,
-    KeyboardEvent: dom.window.KeyboardEvent,
-    requestAnimationFrame: (callback: FrameRequestCallback) => setTimeout(() => callback(Date.now()), 0),
-    cancelAnimationFrame: clearTimeout,
-    getComputedStyle: dom.window.getComputedStyle.bind(dom.window),
-    EventSource: undefined,
-  };
-  for (const [key, value] of Object.entries(globals))
-    Object.defineProperty(globalThis, key, { configurable: true, writable: true, value });
-
-  const row = { id: "s1", threadRef: "web:owner:test", scopeId: "personal:owner", title: "Test" };
-  const entries = [
+  const entries: SessionEntry[] = [
     { seq: 1, type: "user", createdAt: 1, payload: { text: "ship it" } },
     { seq: 2, parentSeq: 1, type: "tool_call", createdAt: 2, payload: { tool: "execute", command: "npm run freeze" } },
     { seq: 3, parentSeq: 2, type: "tool_result", createdAt: 3, payload: { tool: "execute", code: 0, stdout: "ok" } },
@@ -50,52 +14,10 @@ test("tool rows render their label beside a mono detail chip, tinted by outcome"
     { seq: 7, parentSeq: 6, type: "tool_result", createdAt: 7, payload: { tool: "recall" } },
     { seq: 8, type: "assistant", createdAt: 8, payload: { text: "done" } },
   ];
-  globalThis.fetch = async (input) => {
-    const path = String(input);
-    if (path.includes("runtime-config"))
-      return Response.json({
-        scopeId: row.scopeId,
-        approvedHarnesses: ["pi"],
-        modelsByHarness: { pi: ["gpt-5.6-sol"] },
-        modelCatalog: { "gpt-5.6-sol": metadata("gpt-5.6-sol", "GPT-5.6 Sol") },
-        orgDefault: { harnessId: "pi", modelId: "gpt-5.6-sol", revision: 0 },
-        effective: { harnessId: "pi", modelId: "gpt-5.6-sol" },
-        scopeOverride: null,
-      });
-    if (path.includes("/api/runs/active")) return Response.json({ runId: null, queued: [] });
-    if (path.endsWith("/approvals")) return Response.json({ approvals: [] });
-    if (path.startsWith("/api/sessions/s1")) return Response.json({ session: row, entries, earlierEntries: 0 });
-    if (path === "/api/sessions") return Response.json({ sessions: [row] });
-    if (path === "/api/contexts") return Response.json({ contexts: [] });
-    throw new Error(`Unexpected request: ${path}`);
-  };
-  const vite = await createServer({ server: { middlewareMode: true, hmr: false }, appType: "custom" });
-  let conv: Conversation | undefined;
+  const boot = await bootConversation({ entries });
+  const { host } = boot;
   try {
-    await vite.ssrLoadModule("/src/shell.ts");
-    const { appState } = await vite.ssrLoadModule("/src/shell-state.ts");
-    const { sessionsState } = await vite.ssrLoadModule("/src/sessions.ts");
-    const { createConversation } = await vite.ssrLoadModule("/src/conversations.ts");
-    const { entriesToMessages } = await vite.ssrLoadModule("/src/core-bridge.ts");
-    const { transcriptModel } = await vite.ssrLoadModule("/src/model-options.ts");
-    const { seedRuntimeConfig } = await vite.ssrLoadModule("/src/composer.ts");
-    seedRuntimeConfig(row.scopeId, await (await fetch("/api/runtime-config")).json());
-    appState.me = { user: "owner", org: "test" };
-    appState.currentView = "chats";
-    sessionsState.list = [row];
-    const host = document.querySelector<HTMLElement>("#main")!;
-    appState.mainEl = host;
-    conv = createConversation({
-      pane: true,
-      ownsUrl: false,
-      container: () => host,
-      claimContainer: () => host,
-      visible: () => true,
-      density: () => "full",
-      onDensityChange() {},
-      ensureDeliveryStream() {},
-    }) as Conversation;
-    conv.mountContinuable(row.threadRef, row.id, row.scopeId, entriesToMessages(entries, transcriptModel()));
+    boot.mount();
     await until(() => host.querySelectorAll(".tool-row").length === 3);
 
     const ok = host.querySelector(".tool-row.tool-ok")!;
@@ -113,7 +35,6 @@ test("tool rows render their label beside a mono detail chip, tinted by outcome"
     assert.equal(bare.querySelector(".tool-name"), null, "a row without detail shows only its label");
     assert.equal(bare.querySelector(".tool-label")?.textContent, "Searched memory");
   } finally {
-    conv?.dispose();
-    await vite.close();
+    await boot.dispose();
   }
 });
